Split tests according to class or function
- Update pytest command for github action to include all test files
- Use os.path.dirname() to work well with both Linux and Windows
