+++ .github/workflows/test.yaml
@@ -28,7 +28,7 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        python -m pytest tests/test.py
+        python -m pytest tests
 
   build-windows:
     runs-on: windows-latest
@@ -55,4 +55,4 @@ jobs:
           flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Test with pytest
         run: |
-          python -m pytest tests/test.py
+          python -m pytest tests
